Move menus to src root
