+++ lib/src/main/java/com/zetcode/menus/CheckBoxMenuItemEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.menus;
 
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
