+++ lib/src/main/java/com/zetcode/menus/RadioMenuItemEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.menus;
 
 import javax.swing.BorderFactory;
 import javax.swing.ButtonGroup;
