+++ lib/src/main/java/com/zetcode/menus/exit.png
