+++ lib/src/main/java/com/zetcode/menus/exit2.png
