+++ lib/src/main/java/com/zetcode/menus/new.png
