+++ lib/src/main/java/com/zetcode/menus/open.png
