+++ lib/src/main/java/com/zetcode/menus/open2.png
